requirements.txt
python-gitlab==1.4.0
GitPython==2.1.10
